requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0

# Web demo
flask>=2.0.0
plotly>=5.13.0

# Utilities
tqdm>=4.65.0
psutil>=5.9.0
py3nvml>=0.2.7

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0